java/src/main/java/threesumclosest/ThreeSumClosest.java
package threesumclosest;

import java.util.*;

/**
 * Given an integer array nums of length n and an integer target, find three integers in nums such that the sum is closest to target.
 * Return the sum of the three integers.
 * You may assume that each input would have exactly one solution.
 * <p>
 * Example 1:
 * Input: nums = [-1,2,1,-4], target = 1
 * Output: 2
 * Explanation: The sum that is closest to the target is 2. (-1 + 2 + 1 = 2).
 * <p>
 * Example 2:
 * Input: nums = [0,0,0], target = 1
 * Output: 0
 * Explanation: The sum that is closest to the target is 0. (0 + 0 + 0 = 0).
 * <p>
 * Constraints:
 * - 3 <= nums.length <= 500
 * - -1000 <= nums[i] <= 1000
 * - -104 <= target <= 104
 */

public class ThreeSumClosest {
    public int threeSumClosest(int[] nums, int target) {
        Arrays.sort(nums);

        int min = Integer.MAX_VALUE;
        int[] three = new int[3];

        for (int i = 0; i < nums.length - 2; i++) {
            for (int j = i + 1; j < nums.length - 1; j++) {
                for (int k = j + 1; k < nums.length; k++) {
                    if (min > Math.abs(target - (nums[i] + nums[j] + nums[k]))) {
                        min = Math.abs(target - (nums[i] + nums[j] + nums[k]));
                        three[0] = nums[i];
                        three[1] = nums[j];
                        three[2] = nums[k];
                    }
                }
            }
        }
        return three[0] + three[1] + three[2];
    }
}
